+++ SandboxiePlus/SandMan/SbiePlusAPI.cpp
@@ -1098,3 +1098,152 @@ QString CSandBoxPlus::GetFullCommand(const QString& Command)
 	//	FullCmd.insert(1, m_FilePath);
 	return FullCmd.replace("%BoxRoot%", m_FilePath, Qt::CaseInsensitive);
 }
+
+
+///////////////////////////////////////////////////////////////////////////////
+// CSbieTemplatesEx
+//
+
+#include <ntstatus.h>
+#define WIN32_NO_STATUS
+typedef long NTSTATUS;
+#include <Windows.h>
+//#include <Winternl.h>
+
+#include "..\..\Sandboxie\common\win32_ntddk.h"
+
+#define MAX_KEY_NAME 255
+#define MAX_VALUE_NAME 16383
+#define MAX_VALUE_DATA 1024000
+
+#include <comdef.h>
+#include <wuapi.h>
+
+void CSbieTemplatesEx::CollectUpdates()
+{
+	IUpdateSession* updateSession = NULL;
+	IUpdateSearcher* updateSearcher = NULL;
+	ISearchResult* searchResult = NULL;
+	IUpdateCollection* updates = NULL;
+	HRESULT res;
+
+	res = CoInitializeEx(NULL, 0);
+	if (FAILED(res)) {
+		theGUI->OnLogMessage(tr("Failed to initialize COM"));
+		goto cleanup;
+	}
+
+	res = CoCreateInstance(CLSID_UpdateSession, NULL, CLSCTX_INPROC_SERVER, IID_IUpdateSession, (LPVOID*)&updateSession);
+	if (FAILED(res)) {
+		theGUI->OnLogMessage(tr("Failed to create update session"));
+		goto cleanup;
+	}
+
+	res = updateSession->CreateUpdateSearcher(&updateSearcher);
+	if (FAILED(res)) {
+		theGUI->OnLogMessage(tr("Failed to create update searcher"));
+		goto cleanup;
+	}
+
+	res = updateSearcher->put_IncludePotentiallySupersededUpdates(VARIANT_TRUE);
+	if (FAILED(res)) {
+		theGUI->OnLogMessage(tr("Failed to set search options"));
+		goto cleanup;
+	}
+
+	{BSTR criteria = SysAllocString(L"IsInstalled=1"); // or IsHidden=1
+	res = updateSearcher->Search(criteria, &searchResult);
+	SysFreeString(criteria); }
+	if (FAILED(res)) {
+		theGUI->OnLogMessage(tr("Failed to search for updates"));
+		goto cleanup;
+	}
+
+	res = searchResult->get_Updates(&updates);
+	if (FAILED(res)) {
+		theGUI->OnLogMessage(tr("Failed to retrieve update list from search result"));
+		goto cleanup;
+	}
+
+	LONG updateCount;
+	res = updates->get_Count(&updateCount);
+	if (FAILED(res)) {
+		theGUI->OnLogMessage(tr("Failed to get update count"));
+		goto cleanup;
+	}
+
+
+	for (LONG i = 0L; i < updateCount; ++i)
+	{
+		QVariantMap entry;
+
+		IUpdate* update = NULL;
+		res = updates->get_Item(i, &update);
+		if (!FAILED(res))
+		{
+			IStringCollection* updateKBIDs = NULL;
+			res = update->get_KBArticleIDs(&updateKBIDs);
+			if (SUCCEEDED(res)) {
+				LONG kbIDCount;
+				res = updateKBIDs->get_Count(&kbIDCount);
+				if (SUCCEEDED(res)) {
+					QVariantList kb;
+					for (LONG j = 0L; j < kbIDCount; ++j) {
+						BSTR kbID;
+						res = updateKBIDs->get_Item(j, &kbID);
+						if (FAILED(res))
+							continue;
+						kb.append("KB" + QString::fromWCharArray(kbID));
+						SysFreeString(kbID);
+					}
+					entry["kb"] = kb;
+				}
+				updateKBIDs->Release();
+			}
+
+			BSTR updateTitle;
+			res = update->get_Title(&updateTitle);
+			if (!FAILED(res)) {
+				entry["title"] = QString::fromWCharArray(updateTitle);
+				SysFreeString(updateTitle);
+			}
+
+			update->Release();
+		}
+		m_Updates.append(entry);
+	}
+
+cleanup:
+	if (updates != NULL) updates->Release();
+	if (searchResult != NULL) searchResult->Release();
+	if (updateSearcher != NULL) updateSearcher->Release();
+	if (updateSession != NULL) updateSession->Release();
+
+	CoUninitialize();
+}
+
+void CSbieTemplatesEx::Reset()
+{
+	CSbieTemplates::Reset();
+	m_Updates.clear();
+}
+
+QList<QVariantMap> CSbieTemplatesEx::GetUpdates()
+{
+	if (m_Updates.isEmpty())
+		CollectUpdates();
+
+	return m_Updates;
+}
+
+bool CSbieTemplatesEx::CheckUpdates(const QString& value)
+{
+	if (m_Updates.isEmpty())
+		CollectUpdates();
+
+	auto I = std::find_if(m_Updates.begin(), m_Updates.end(), [value](const QVariantMap& cur)->bool {
+		return cur["kb"].toStringList().contains(value, Qt::CaseInsensitive);
+	});
+	bool bRet = (I != m_Updates.end());
+	return bRet;
+}